Improve linux module installer
Description
===========

Second patch to gain Unity Hub parity for module installation.
Linux had already all nessary installer types covered. This patch
concentrates more on cleanup and code sharing. I moved a few functions
with the same logic into a `shared` module just to reduce overhead. I
also retouched the general logic in the `_install_module` function to
make it similar to the macos variant.

Changes
=======

![IMPROVE] ![LINUX] module installer logic
![ADD] shared installer module for common functions
