+++ uvm_core/src/sys/mod.rs
@@ -1,3 +1,4 @@
+pub mod shared;
 cfg_if::cfg_if! {
     if #[cfg(target_os = "macos")] {
         mod mac;
