+++ uvm_core/src/sys/shared/mod.rs
@@ -1,0 +1,1 @@
+pub mod installer;
